requirements.txt
Flask==3.0.2
requests==2.31.0
slack-sdk==3.27.0
google-auth==2.28.1
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0
numpy==1.26.4
boto3==1.34.39
python-dotenv==1.0.1
jira==3.6.0
